Combination of edits to CP4 data management and bug fixes to tracking.
Note that not all CP4 data is included and some settings might be weird in the regridding files in this commit.
Includes updates to required modules, using lon or longitude as a key for longitude while reading files, fixing
option for more than one argument to the AEW_TRACKING_CODE function.